fix: set height more reliably
If the user has autocmds that potentially resize windows on
BufEnter/WinEnter (like I do), those can trigger when we do `lopen` or
`copen`, which is a problem because that happens *before* we can set the
default `winfixheight` option. To work around this, we call the
`lopen/copen` command twice. The second call will enforce the height of
the window *after* all the default options are set.